BUGFIX.
Fixed small bug with displaying trades.

+++ src/main/java/org/kohera/metctools/portfolio/Trade.java
@@ -239,7 +239,7 @@ public class Trade {
 				getSymbol(),
 				getLastPrice().floatValue(),
 				(side==Side.BUY?"+":(side==Side.SELL?"-":"")),
-				quantity,
+				quantity.intValue(),
 				costBasis.floatValue());
 	}
